Update dependencies, rename colors.android-colors.xml to colors.android.xml
