+++ __tests__/colors.test.js
@@ -9,7 +9,7 @@ const colorsAseJSON = require('../dist/colors.ase.json');
 const colorsRawJSON = require('../dist/colors.raw.json');
 
 const colorFiles = [
-  'colors.android-colors.xml',
+  'colors.android.xml',
   'colors.color-map.scss',
   'colors.custom-properties.css',
   'colors.scss',
