+++ package.json
@@ -63,34 +63,34 @@
   "devDependencies": {
     "ase-util": "^1.0.1",
     "ase-utils": "^0.1.1",
-    "browser-sync": "^2.24.5",
+    "browser-sync": "^2.24.6",
     "dashify": "^1.0.0",
-    "eslint": "^5.0.1",
+    "eslint": "^5.2.0",
     "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-prettier": "^2.6.1",
-    "eslint-plugin-shopify": "^22.1.0",
-    "fs-extra": "^6.0.1",
+    "eslint-plugin-prettier": "^2.6.2",
+    "eslint-plugin-shopify": "^23.0.0",
+    "fs-extra": "^7.0.0",
     "gulp": "^3.9.1",
     "gulp-load-plugins": "^1.5.0",
     "gulp-plumber": "^1.2.0",
-    "gulp-rename": "^1.3.0",
+    "gulp-rename": "^1.4.0",
     "gulp-restart": "^0.1.1",
     "gulp-sass": "^4.0.1",
     "gulp-sourcemaps": "^2.6.4",
     "gulp-theo": "^2.0.0",
     "http-server": "^0.11.1",
     "immutable": "^3.8.2",
-    "jest": "^23.2.0",
+    "jest": "^23.4.1",
     "js-yaml": "^3.12.0",
     "lodash": "^4.17.10",
-    "node-fetch": "^2.1.2",
-    "nodemon": "^1.17.5",
+    "node-fetch": "^2.2.0",
+    "nodemon": "^1.18.3",
     "npm-run-all": "^4.1.3",
     "prettier": "1.13.7",
-    "renamer": "^0.7.3",
+    "renamer": "^1.0.0",
     "rimraf": "^2.6.2",
     "run-sequence": "^2.2.1",
-    "theo": "^7.0.1",
+    "theo": "^8.0.0-beta.2",
     "tinycolor2": "^1.4.1",
     "xml": "^1.0.1",
     "yaml-lint": "^1.2.4"
